Split CA tests into three executables and update all documentation
