+++ source/tests/unit/CMakeLists.txt
@@ -57,6 +57,16 @@ genesys_add_unit_test(genesys_test_cellular_automata_neighborhood
     genesys_plugins_components_minimal
 )
 
+genesys_add_unit_test(genesys_test_cellular_automata_nonuniform
+    test_cellular_automata_nonuniform.cpp
+    genesys_plugins_components_minimal
+)
+
+genesys_add_unit_test(genesys_test_cellular_automata_boundary
+    test_cellular_automata_boundary.cpp
+    genesys_plugins_components_minimal
+)
+
 genesys_add_unit_test(genesys_test_support_modelmanager
     test_support_modelmanager.cpp
     genesys_kernel_simulator_support
@@ -272,6 +282,8 @@ add_custom_target(genesys_kernel_unit_tests
         genesys_test_simulator_runtime
         genesys_test_parser_expressions
         genesys_test_cellular_automata_neighborhood
+        genesys_test_cellular_automata_nonuniform
+        genesys_test_cellular_automata_boundary
         genesys_test_support_modelmanager
         genesys_test_support_persistence
         genesys_test_runtime_pluginmanager
@@ -351,6 +363,8 @@ add_custom_target(genesys_kernel_unit_tests_run
     COMMAND $<TARGET_FILE:genesys_test_simulator_runtime>
     COMMAND $<TARGET_FILE:genesys_test_parser_expressions>
     COMMAND $<TARGET_FILE:genesys_test_cellular_automata_neighborhood>
+    COMMAND $<TARGET_FILE:genesys_test_cellular_automata_nonuniform>
+    COMMAND $<TARGET_FILE:genesys_test_cellular_automata_boundary>
     COMMAND $<TARGET_FILE:genesys_test_support_modelmanager>
     COMMAND $<TARGET_FILE:genesys_test_support_persistence>
     COMMAND $<TARGET_FILE:genesys_test_runtime_pluginmanager>
